feat: Rough draft of file naming convention and switch clap framework
I now have a rough but non-functional outline of how changing file
naming conventions is going to work. Definitely need to work on the
safety though. Also now using the derive feature in clap instead of the
other app thing. I think it will be easier to manage and is definitely
more familiar to me since I use it all the time for my other Rust CLI
tools.